Header, js, font, and formatting
I fixed the fonts, added a new css to link to all the forms, ill mainstream it better later but i made it so every single html is connected to the JS and css. i also fixed the header since it would widen the page when we used the old one.